Conflicts: main reformatted src/bin/ic/bench.rs and
src/cryptanalysis/ic_boundary.rs under its new lint gate, around code
this branch had replaced (the matched rho references of ledger §18);
this branch's logic is kept.  In koblitz_fast.rs both sides added items
after FrobeniusCanon (FrobeniusPowers here, the AVX-512 least rotation
there); both are kept.

main now requires every changed .rs file to be rustfmt-clean and the
crate to pass `cargo clippy --all-targets -- -D warnings` on 1.98, so the
seven files this branch changes are formatted and three findings in
src/bin/ic/rho.rs are fixed (is_multiple_of, is_none_or, a late
initialisation).

Checked on the merge: clippy -D warnings clean on 1.98; the refactored
walk replays all 3,072 runs of the frozen §18 ladder and the first
batches of the frozen §19 k0n41 run exactly; ic_boundary and
koblitz_fast unit tests and the ic_framework / ic_progress suites pass.

+++ src/cryptanalysis/modular_polynomial.rs
@@ -104,7 +104,7 @@ impl ModularPolynomial {
         self.coeffs
             .iter()
             .filter(|(_, c)| !c.is_zero())
-            .map(|(_, c)| c.bits() as u64)
+            .map(|(_, c)| c.bits())
             .sum()
     }
 }
